// frontend/src/types/event-monetization.ts
export enum EventCategory {
    CONCERT = 'concert',
    FESTIVAL = 'festival',
    WORKSHOP = 'workshop',
    SPORT = 'sport',
    RESTAURANT = 'restaurant',
    NIGHTLIFE = 'nightlife',
    CULTURE = 'culture',
    OTHER = 'other'
}

export enum AdTier {
    BRONZE = 'BRONZE',
    SILVER = 'SILVER',
    GOLD = 'GOLD'
}

export enum AdSize {
    SMALL = 'SMALL',
    MEDIUM = 'MEDIUM',
    LARGE = 'LARGE'
}